feat(GameView):  Create initial Game View
Install react-router-dom. Hook game view to lobby. Hook game board to game view.
