+++ app/components/Lobby.jsx
@@ -1,23 +1,8 @@
 import React, { Component } from 'react'
 import firebase from 'APP/fire'
+import { browserHistory } from 'react-router'
 
 export default class Lobby extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      allWords: []
-    }
-    this.onClickNewGame = this.onClickNewGame.bind(this)
-  }
-
-  componentDidMount() {
-    const allWords = firebase.database().ref('words/')
-    allWords.on('value', snapshot => {
-      const wordArray = snapshot.val()
-      this.setState({allWords: wordArray})
-    })
-  }
-
   // CREATES A NEW GAME INSTANCE AND HAS ACCESS TO THE KEY
   onClickNewGame(event, data) {
     event.preventDefault()
@@ -26,7 +11,7 @@ export default class Lobby extends Component {
     }
     const newGameRef = firebase.database().ref('/gameInstances').push(newGameInstance)
     const newGameKey = newGameRef.key
-    firebase.database().ref(`/gameInstances/${newGameKey}`).update({gameId: newGameKey})
+    newGameRef.then(() => browserHistory.push(`/game-view/${newGameKey}/wordassassins`))
   }
 
   render() {
